+++ flexmeasures/conftest.py
@@ -148,11 +148,36 @@ def create_test_accounts(db) -> dict[str, Account]:
         ],
     )
     db.session.add(multi_role_account)
+    consultant_account_role = AccountRole(
+        name="Consultant", description="A Consultant Account"
+    )
+    # Create Consultant and ConsultantClient account.
+    # The ConsultantClient account needs the account id of the Consultant account so the order is important.
+    consultant_account = Account(
+        name="Test Consultant Account", account_roles=[consultant_account_role]
+    )
+    db.session.add(consultant_account)
+    consultant_client_account_role = AccountRole(
+        name="ConsultantClient",
+        description="A Client of a Consultant",
+    )
+    consultant_account_id = (
+        Account.query.filter_by(name="Test Consultant Account").one_or_none().id
+    )
+    consultant_client_account = Account(
+        name="Test ConsultantClient Account",
+        account_roles=[consultant_client_account_role],
+        consultancy_account_id=consultant_account_id,
+    )
+    db.session.add(consultant_client_account)
     return dict(
         Prosumer=prosumer_account,
         Supplier=supplier_account,
         Dummy=dummy_account,
+        Empty=empty_account,
         Multi=multi_role_account,
+        Consultant=consultant_account,
+        ConsultantClient=consultant_client_account,
     )
 
 
@@ -219,6 +244,23 @@ def create_roles_users(db, test_accounts) -> dict[str, User]:
             user_roles=dict(
                 name=ADMIN_ROLE, description="A user who can do everything."
             ),
+        )
+    )
+    new_users.append(
+        create_user(
+            username="Test Consultant User",
+            email="test_consultant_user@seita.nl",
+            account_name=test_accounts["Consultant"].name,
+            password="testtest",
+            user_roles=dict(name="customer-manager"),
+        )
+    )
+    new_users.append(
+        create_user(
+            username="Test Consultant User without customer-manager role",
+            email="test_consultant_user_without_customer_manager_access@seita.nl",
+            account_name=test_accounts["Consultant"].name,
+            password="testtest",
         )
     )
     return {user.username: user.id for user in new_users}
@@ -323,11 +365,18 @@ def create_generic_assets(
         owner=setup_accounts["Supplier"],
     )
     db.session.add(test_wind_turbine)
+    test_consultant_client_asset = GenericAsset(
+        name="Test ConsultantClient Asset",
+        generic_asset_type=setup_generic_asset_types["wind"],
+        owner=setup_accounts["ConsultantClient"],
+    )
+    db.session.add(test_consultant_client_asset)
 
     return dict(
         troposphere=troposphere,
         test_battery=test_battery,
         test_wind_turbine=test_wind_turbine,
+        test_consultant_client_asset=test_consultant_client_asset,
     )